adding mandatory notifications
