+++ .gitignore
@@ -105,3 +105,6 @@ dist
 
 # IDE
 .idea
+
+# Redis
+dump.rdb
