--- redis.conf
@@ -1,2 +1,0 @@
-appendonly no
-save ""
